fix: location set 로직을 home 모드에서만 작동하도록 변경
